Use traits for InetAddress and RackResolver util classes
