+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/scheduler/cluster/kubernetes/KubernetesTaskSchedulerImpl.scala
@@ -16,12 +16,6 @@
  */
 package org.apache.spark.scheduler.cluster.kubernetes
 
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.CommonConfigurationKeysPublic
-import org.apache.hadoop.net.{NetworkTopology, ScriptBasedMapping, TableMapping}
-import org.apache.hadoop.yarn.util.RackResolver
-import org.apache.log4j.{Level, Logger}
-
 import org.apache.spark.scheduler.{SchedulerBackend, TaskSchedulerImpl, TaskSet, TaskSetManager}
 import org.apache.spark.util.Utils
 import org.apache.spark.SparkContext
@@ -29,11 +23,11 @@ import org.apache.spark.SparkContext
 private[spark] class KubernetesTaskSchedulerImpl(
     sc: SparkContext,
     rackResolverUtil: RackResolverUtil,
-    inetAddressUtil: InetAddressUtil = new InetAddressUtil) extends TaskSchedulerImpl(sc) {
+    inetAddressUtil: InetAddressUtil = InetAddressUtilImpl) extends TaskSchedulerImpl(sc) {
 
   var kubernetesSchedulerBackend: KubernetesClusterSchedulerBackend = null
 
-  def this(sc: SparkContext) = this(sc, new RackResolverUtil(sc.hadoopConfiguration))
+  def this(sc: SparkContext) = this(sc, new RackResolverUtilImpl(sc.hadoopConfiguration))
 
   override def initialize(backend: SchedulerBackend): Unit = {
     super.initialize(backend)
@@ -71,38 +65,5 @@ private[spark] class KubernetesTaskSchedulerImpl(
           })
         }
       ).getOrElse(rackResolverUtil.resolveRack(hadoopConfiguration, host))
-  }
-}
-
-private[kubernetes] class RackResolverUtil(hadoopConfiguration: Configuration) {
-
-  val scriptPlugin : String = classOf[ScriptBasedMapping].getCanonicalName
-  val tablePlugin : String = classOf[TableMapping].getCanonicalName
-  val isConfigured : Boolean = checkConfigured(hadoopConfiguration)
-
-  // RackResolver logs an INFO message whenever it resolves a rack, which is way too often.
-  if (Logger.getLogger(classOf[RackResolver]).getLevel == null) {
-    Logger.getLogger(classOf[RackResolver]).setLevel(Level.WARN)
-  }
-
-  def checkConfigured(hadoopConfiguration: Configuration): Boolean = {
-    val plugin = hadoopConfiguration.get(
-      CommonConfigurationKeysPublic.NET_TOPOLOGY_NODE_SWITCH_MAPPING_IMPL_KEY, scriptPlugin)
-    val scriptName = hadoopConfiguration.get(
-      CommonConfigurationKeysPublic.NET_TOPOLOGY_SCRIPT_FILE_NAME_KEY, "")
-    val tableName = hadoopConfiguration.get(
-      CommonConfigurationKeysPublic.NET_TOPOLOGY_TABLE_MAPPING_FILE_KEY, "")
-    plugin == scriptPlugin && scriptName.nonEmpty ||
-      plugin == tablePlugin && tableName.nonEmpty ||
-      plugin != scriptPlugin && plugin != tablePlugin
-  }
-
-  def resolveRack(hadoopConfiguration: Configuration, host: String): Option[String] = {
-    val rack = Option(RackResolver.resolve(hadoopConfiguration, host).getNetworkLocation)
-    if (rack.nonEmpty && rack.get != NetworkTopology.DEFAULT_RACK) {
-      rack
-    } else {
-      None
-    }
   }
 }
